rename ProductionInfrastructure -> infrastructure
